Reorganize utility classes
